src_GIP/src_2022_07_19_GIP146/requirements.txt
absl-py==0.12.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
efficientnet-pytorch==0.6.3
future==0.18.2
google-auth==1.30.0
google-auth-oauthlib==0.4.4
grpcio==1.37.1
idna==2.10
imageio==2.9.0
importlib-metadata==4.0.1
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
munch==2.5.0
networkx==2.5.1
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.3.56
pandas==1.1.5
Pillow==7.2.0
pretrainedmodels==0.7.4
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.17.2
scipy==1.1.0
seaborn==0.11.1
segmentation-models-pytorch==0.2.0
six==1.16.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
tifffile==2020.9.3
timm==0.4.12
torch==1.8.0+cu101
torchvision==0.9.0+cu101
tqdm==4.50.2
typing-extensions==3.10.0.0
urllib3==1.26.4
Werkzeug==2.0.0
zipp==3.4.1
